+++ src/test/java/fi/metatavu/keycloak/scim/server/RealmUserListTestsIT.java
@@ -17,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.*;
  * Tests for SCIM 2.0 User list endpoint
  */
 @Testcontainers
-class UserListTestsIT extends AbstractScimTest {
+class RealmUserListTestsIT extends AbstractRealmScimTest {
 
   @Container
   protected static final KeycloakContainer keycloakContainer = new KeycloakContainer("quay.io/keycloak/keycloak:26.1.2")
